+++ AlfrescoKit/AlfrescoKit/Constants/AKConstants.m
@@ -21,7 +21,7 @@
 #import "AKConstants.h"
 
 // Max items
-NSUInteger const kMaximumItemsToRetrieveAtOneTime = 50;
+int const kMaximumItemsToRetrieveAtOneTime = 50;
 
 // AlfrescoKit bundle name
 NSString * const kAlfrescoKitBundleName = @"AlfrescoKitBundle";
